Add iterator around Log.append
Adds an iterator around Log.append that iterates over the MemoryRecords buffer that is passed to append and passes individual batches to the inner append call. This allows each batch to be processed in isolation from each other, which means they are considered separately for whether or not offsets need to be reassigned to a batch. For a producer that can produce multiple batches for a single topic-partition in a single Produce request, this is much more efficient for appending those batches.

RB=1216872