HOTFIX Eliminate reflection in index object creation
We identified that we spend a lot of time in reflection when creating
OffsetIndex, TimeIndex, or other implementations of
AbstractIndex[K, V], because of the Logging mixin. When the broker is
bootstrapping it's just doing this in a tight loop, so this time adds
up.

This patch moves the logging to a companion object, statically
initializing the logger.

This isn't a perfect fix, since this actually changes the logger names
for several types. For example kafka.log.OffsetIndex and
kafka.log.TimeIndex now have the logger name
kafka.log.AbstractIndex$. There are a couple of different ways to fix
this, but not using this Logging mixin.

I'm opting to go the route of minimal change for now rather than
trying to preserve the logger names, as I'd like to keep the runtime
cost as low as possible and we're short on time. I don't actually
think the logger name change will affect us internally. Upstream will
care though, so I'll redo this in a way that preserves the logger
names later.

RB=1225180
BUG=LIKAFKA-14964
G=Kafka-Code-Reviews
A=rrosenbl,jqin