[LI-HOTFIX] Clean up inFlightBatches in Sender when ProducerBatch split happens (#34)
TICKET = KAFKA-8202
LI_DESCRIPTION =
As part of the producer delivery timeout support (KAFKA-5886, KIP-91),
the inFlightBatches map is introduced in the producer Sender but
currently the batches in that map will only be cleaned up when
completing, expiring and failing a batch. However, when batch split
happens, the original batch (created prior to the split) will never be
cleaned up from inFlightBatches map. This cause excessive memory
pressure in the producer because these zombie batches are normally big
batches and eventully it will cause OutOfMemory in the producer.

This patch addresses the issue by removing the original batch from the
Sender inFlightBatchMap after split happens. The test case in SenderTest
also gets modified accordingly to cover this scenario.

EXIT_CRITERIA = TICKET [KAFKA-8202]